backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
discord.py==2.4.0
apscheduler==3.10.4
numpy>=1.26.4
pandas>=2.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
